Moves ConfirmationPopUp to a utils folder easy re-use
